DSTS-21 Design ticket creation form UI
